+++ crates/acts/src/scheduler/process/process.rs
@@ -11,7 +11,6 @@ use crate::{
     utils::{self, consts},
 };
 use serde::Deserialize;
-use std::collections::HashMap;
 use std::{
     cell::RefCell,
     fmt,