+++ crates/acts/src/scheduler/process/task/workflow.rs
@@ -14,7 +14,6 @@ impl ActTask for Workflow {
             });
         }
 
-
         Ok(())
     }